+++ test/protractor/basePage.js
@@ -196,12 +196,26 @@ function BasePage() {
    * @param text the text with which to autocomplete
    */
   this.autocompPickFirst = function(field, text) {
+    this.autocompPickNth(field, text, 1);
+  };
+
+
+  /**
+   *  Sets the field's value to the given text string, and picks the nth
+   *  autocompletion result.
+   * @param field the autocompleting field
+   * @param text the text with which to autocomplete.  This can be null or the
+   *  empty string for prefetch lists.
+   * @param n the number of the list item to pick (starting at 1).
+   */
+  this.autocompPickNth = function(field, text, n) {
     this.clearField(field);
     expect(field.getAttribute('value')).toEqual('');
     field.click();
-    field.sendKeys(text);
+    if (text)
+      field.sendKeys(text);
     this.waitForSearchResults();
-    this.firstSearchRes.click();
+    this.searchResult(n).click();
   };
 
 
@@ -227,6 +241,38 @@ function BasePage() {
     });
   }
 
+
+  /**
+   *  Outputs the browser's console messages.
+   */
+  this.printBrowserConsole = function() {
+    browser.manage().logs().get('browser').then(function(browserLogs) {
+      if (browserLogs.length > 0) {
+        console.log("Messages from browser's console");
+        browserLogs.forEach(function(log){
+          console.log(log.message);
+        });
+        console.log("End of messages from browser's console");
+      }
+    });
+  }
+
+
+  /**
+   *  Sends key events to a field.
+   * @param field the field to which the characters should be send
+   * @param text the text whose characters should be sent
+   */
+  this.sendKeys = function(field, text) {
+    // For some reason, on RHEL 7 in Chrome, if you focus a field and then send
+    // key events withouht sleeping a bit between the two, the autocompleter's
+    // getUpdates function does not get called before the field value is updated
+    // with the last character.  The problem can also be solved by increasing
+    // the autocompleter options.frequency setting, but it only happens when
+    // testing, so I don't want to make that slower.
+    browser.sleep(100);
+    field.sendKeys(text);
+  }
 };
 
 module.exports = {BasePage: BasePage};