+++ test/protractor/conf.js
@@ -79,7 +79,10 @@ exports.config = {
     var SpecReporter = require('jasmine-spec-reporter');
     // add jasmine spec reporter
     jasmine.getEnv().addReporter(new SpecReporter({displayStacktrace: true}));
+  },
 
-
+  onComplete: () => {
+    var basePage = new (require('./basePage').BasePage)();
+    basePage.printBrowserConsole();
   }
 };