+++ CHANGELOG.md
@@ -3,6 +3,23 @@
 This log documents significant changes for each release.  This project follows
 [Semantic Versioning](http://semver.org/).
 
+## [15.0.0] - 2017-07-07
+### Added
+- A new function, getSelectedItemData(), is now available for retrieving all
+  information about the selected list items.
+### Changed
+- The expected data format for search autocompleters has changed slightly.  In
+  particular, the optional fourth element of the the JSON data returned by the
+  AJAX call is now an array of code system names declaring the code system for
+  each of the codes in the codes array in position 1 of the JSON data.
+  Previously, the optional fourth parameter was a boolean indicating that the
+  display string data (position 3) might have span tags for formatting.
+  However, it has been some time since the code used this parameter.  Span tags
+  are still permitted in display strings, but it is no longer important that the
+  autocompleter be aware of that.
+- The parameter "tokens" has been renamed to "wordBoundaryChars", though for now
+  the older parameter name will still work as well.
+
 ## [14.0.0] - 2017-05-25
 ### Changed
 - The AngularJS directive for search lists now places the "extra data" fields
@@ -20,7 +37,7 @@ This log documents significant changes for each release.  This project follows
 ### Fixed
 - Fixed an accessibility issue with search lists by adding code to suggest on
   field focus that users start typing to see a list of matching results.
- 
+
 ## [13.0.0] - 2017-05-10
 ### Fixed
 - For AngularJS directive single-select lists, clearing a single-select list now
@@ -70,7 +87,7 @@ This log documents significant changes for each release.  This project follows
 
 ## [10.2.0] - 2017-03-14
 ### Changed
-- Replaced the search field "list" icon with a magnifying glass icon. 
+- Replaced the search field "list" icon with a magnifying glass icon.
 
 ## [10.1.1] - 2017-02-01
 ### Fixed
@@ -161,7 +178,7 @@ This log documents significant changes for each release.  This project follows
   querying.  This solves a problem in which a selected list item could look
   invalid if a query was sent for the combined field value of a multi-field list.
 - The AngularJS directive now clears a search field's result cache if its URL
-  changes. 
+  changes.
 
 ## [9.0.0] - 2016-06-07
 ### Fixed
@@ -178,7 +195,7 @@ This log documents significant changes for each release.  This project follows
   https://github.com/angular/protractor/issues/3200).
 
 ## [8.1.0] - 2016-04-07
-### Added 
+### Added
 - Search autocompleters now have a "setURL" method for updating the URL which
   also takes care of clearing the cached autocompletion results (from the
   previous URL), as well as a clearCachedResults method for clearing the cache